refactoring - fixed formatting of dayview on homepage - moved api calls to lib/api/ - moved calendarProps to calendar component
