+++ src/app/auth/login/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from 'next/server';
-import { msalInstance } from '@/lib/msal';
+import { msalInstance } from '@/lib/api/msal';
 
 /**
  * Redirects the user to the Microsoft SSO login page.
